+++ .test-database/pgsql/1.create-some-stuff.sql
@@ -1,7 +1,9 @@
+-- @conn PGSQL
 -- Extracted from w3schools
 
--- @name create company table
-CREATE TABLE COMPANY(
+-- @block create company table
+-- @conn PGSQL
+CREATE TABLE IF NOT EXISTS COMPANY(
     ID INT PRIMARY KEY NOT NULL,
     NAME TEXT NOT NULL,
     AGE INT NOT NULL,
@@ -9,8 +11,8 @@ CREATE TABLE COMPANY(
     SALARY REAL
 );
 
--- @name create department table
-CREATE TABLE DEPARTMENT(
+-- @block create department table
+CREATE TABLE IF NOT EXISTS DEPARTMENT(
     ID INT PRIMARY KEY NOT NULL,
     DEPT CHAR(50) NOT NULL,
     EMP_ID INT NOT NULL