+++ packages/core/dialect/sqlite/queries.ts
@@ -1,10 +1,11 @@
 import { DialectQueries } from '@sqltools/core/interface';
+import { TREE_SEP } from '../../constants';
 
 export default {
   describeTable: 'SELECT * FROM pragma_table_info(\':table\') ORDER BY cid ASC',
   fetchColumns: `
 SELECT
-  M.type || 's' || '/' || M.name || '/' || C.name AS tree,
+  M.type || 's' || '${TREE_SEP}' || M.name || '${TREE_SEP}' || C.name AS tree,
   C.*
 FROM
   pragma_table_info(':table') AS C
@@ -15,7 +16,7 @@ ORDER BY
   fetchTables: `SELECT
       name AS tableName,
       type,
-      type || 's' || '/' || name AS tree
+      type || 's' || '${TREE_SEP}' || name AS tree
     FROM
       sqlite_master
     WHERE type = 'table' OR type = 'view'