+++ packages/plugins/connection-manager/language-server.ts
@@ -9,6 +9,7 @@ import { ConnectionDataUpdatedRequest, ConnectRequest, DisconnectRequest, GetCon
 import actions from './store/actions';
 import DependencyManager from '../dependency-manager/language-server';
 import { DependeciesAreBeingInstalledNotification } from '../dependency-manager/contracts';
+import { decorateException } from '@sqltools/core/utils/errors';
 
 export default class ConnectionManagerPlugin implements SQLTools.LanguageServerPlugin {
   private server: SQLTools.LanguageServerInterface;
@@ -114,7 +115,7 @@ export default class ConnectionManagerPlugin implements SQLTools.LanguageServerP
     conn: ConnectionInterface;
     password?: string;
   }): Promise<ConnectionInterface> => {
-    if (!req.conn) {
+    if (!req || !req.conn) {
       return undefined;
     }
     let c = this.getConnectionInstance(req.conn) || new Connection(req.conn, this.server.telemetry);
@@ -128,12 +129,16 @@ export default class ConnectionManagerPlugin implements SQLTools.LanguageServerP
         return c.serialize();
       })
       .catch(e => {
+        e = decorateException(e, { conn: req.conn });
         if (e.data && e.data.notification) {
-          if (DependencyManager.runningJobs.includes(e.data.args.conn.dialect)) {
+          if (req.conn.dialect && DependencyManager.runningJobs.includes(req.conn.dialect)) {
             return void this.server.sendNotification(DependeciesAreBeingInstalledNotification, e.data.args);
           }
           return void this.server.sendNotification(e.data.notification, e.data.args);
         }
+
+        this.server.telemetry.registerException(e);
+
         throw e;
       });
   };