+++ packages/plugins/connection-manager/screens/settings.ts
@@ -13,9 +13,24 @@ export default class SettingsWebview extends WebviewProvider {
       switch (action) {
         case 'createConnection':
           return this.createConnection(payload);
+        case 'updateConnection':
+          return this.updateConnection(payload);
         default:
         break;
       }
+    });
+  }
+
+  private updateConnection = async ({ connInfo, isGlobal, editId }) => {
+    console.log({ connInfo, isGlobal, editId });
+    commands.executeCommand(`${EXT_NAME}.updateConnection`, editId, connInfo, isGlobal ? 'Global' : undefined)
+    .then(() => {
+      this.postMessage({ action: 'updateConnectionSuccess', payload: { isGlobal, connInfo: { ...connInfo, id: getConnectionId(connInfo) } } });
+    }, (payload = {}) => {
+        payload = {
+          message: (payload.message || payload || '').toString(),
+        }
+        this.postMessage({ action: 'updateConnectionError', payload });
     });
   }